add SDK test and check Result in RPC response
