+++ pkg/engine/lifecycletest/pulumi_test.go
@@ -5498,243 +5498,6 @@ func TestConstructCallDependencyDedeuplication(t *testing.T) {
 	})
 }
 
-func TestDestroyContinueOnError(t *testing.T) {
-	t.Parallel()
-	loaders := []*deploytest.ProviderLoader{
-		deploytest.NewProviderLoader("pkgA", semver.MustParse("1.0.0"), func() (plugin.Provider, error) {
-			return &deploytest.Provider{}, nil
-		}, deploytest.WithoutGrpc),
-		deploytest.NewProviderLoader("pkgB", semver.MustParse("1.0.0"), func() (plugin.Provider, error) {
-			return &deploytest.Provider{
-				DeleteF: func(urn resource.URN, id resource.ID, oldInputs, oldOutputs resource.PropertyMap,
-					timeout float64,
-				) (resource.Status, error) {
-					return resource.StatusOK, errors.New("intentionally failed delete")
-				},
-			}, nil
-		}, deploytest.WithoutGrpc),
-	}
-
-	createResource := true
-
-	programF := deploytest.NewLanguageRuntimeF(func(_ plugin.RunInfo, monitor *deploytest.ResourceMonitor) error {
-		if createResource {
-			resp, err := monitor.RegisterResource("pkgA:m:typA", "unrelated1", true, deploytest.ResourceOptions{})
-			assert.NoError(t, err)
-			_, err = monitor.RegisterResource("pkgA:m:typA", "unrelated2", true, deploytest.ResourceOptions{
-				Dependencies: []resource.URN{resp.URN},
-			})
-			assert.NoError(t, err)
-
-			resp, err = monitor.RegisterResource("pkgA:m:typA", "dependency", true, deploytest.ResourceOptions{})
-			assert.NoError(t, err)
-
-			_, err = monitor.RegisterResource("pkgB:m:typB", "failing", true, deploytest.ResourceOptions{
-				Dependencies: []resource.URN{resp.URN},
-			})
-			assert.NoError(t, err)
-
-			_, err = monitor.RegisterResource("pkgA:m:typA", "anotherUnrelatedRes", true, deploytest.ResourceOptions{})
-			assert.NoError(t, err)
-
-		}
-
-		return nil
-	})
-	hostF := deploytest.NewPluginHostF(nil, nil, programF, loaders...)
-
-	p := &TestPlan{
-		Options: TestUpdateOptions{
-			UpdateOptions: UpdateOptions{
-				ContinueOnError: true,
-			},
-			HostF: hostF,
-		},
-	}
-
-	project := p.GetProject()
-
-	// Run an update to create the resource
-	snap, err := TestOp(Update).Run(project, p.GetTarget(t, nil), p.Options, false, p.BackendClient, nil)
-	require.NoError(t, err)
-	assert.NotNil(t, snap)
-	assert.Len(t, snap.Resources, 7) // We expect 5 resources + 2 providers
-
-	createResource = false
-	snap, err = TestOp(Update).Run(project, p.GetTarget(t, snap), p.Options, false, p.BackendClient, nil)
-	assert.ErrorContains(t, err, "intentionally failed delete")
-	assert.NotNil(t, snap)
-	assert.Len(t, snap.Resources, 4) // We expect 2 resources + 2 providers
-	assert.Equal(t, resource.URN("urn:pulumi:test::test::pulumi:providers:pkgA::default"), snap.Resources[0].URN)
-	assert.Equal(t, resource.URN("urn:pulumi:test::test::pkgA:m:typA::dependency"), snap.Resources[1].URN)
-	assert.Equal(t, resource.URN("urn:pulumi:test::test::pulumi:providers:pkgB::default"), snap.Resources[2].URN)
-	assert.Equal(t, resource.URN("urn:pulumi:test::test::pkgB:m:typB::failing"), snap.Resources[3].URN)
-}
-
-func TestUpContinueOnErrorCreate(t *testing.T) {
-	t.Parallel()
-
-	loaders := []*deploytest.ProviderLoader{
-		deploytest.NewProviderLoader("pkgA", semver.MustParse("1.0.0"), func() (plugin.Provider, error) {
-			return &deploytest.Provider{}, nil
-		}, deploytest.WithoutGrpc),
-		deploytest.NewProviderLoader("pkgB", semver.MustParse("1.0.0"), func() (plugin.Provider, error) {
-			return &deploytest.Provider{
-				CreateF: func(urn resource.URN, news resource.PropertyMap, timeout float64,
-					preview bool,
-				) (resource.ID, resource.PropertyMap, resource.Status, error) {
-					return "", nil, resource.StatusOK, errors.New("intentionally failed create")
-				},
-			}, nil
-		}, deploytest.WithoutGrpc),
-	}
-
-	programF := deploytest.NewLanguageRuntimeF(func(_ plugin.RunInfo, monitor *deploytest.ResourceMonitor) error {
-		failingResp, err := monitor.RegisterResource("pkgB:m:typB", "failing", true, deploytest.ResourceOptions{
-			SupportsResultReporting: true,
-		})
-		assert.NoError(t, err)
-
-		respIndependent1, err := monitor.RegisterResource(
-			"pkgA:m:typA", "independent1", true, deploytest.ResourceOptions{SupportsResultReporting: true})
-		assert.NoError(t, err)
-		respIndependent2, err := monitor.RegisterResource(
-			"pkgA:m:typA", "independent2", true, deploytest.ResourceOptions{
-				SupportsResultReporting: true,
-				Dependencies:            []resource.URN{respIndependent1.URN},
-			})
-		assert.NoError(t, err)
-		_, err = monitor.RegisterResource("pkgA:m:typA", "independent3", true, deploytest.ResourceOptions{
-			SupportsResultReporting: true,
-			Dependencies:            []resource.URN{respIndependent2.URN},
-		})
-		assert.NoError(t, err)
-
-		_, err = monitor.RegisterResource("pkgA:m:typA", "dependentOnFailing", true, deploytest.ResourceOptions{
-			SupportsResultReporting: true,
-			Dependencies:            []resource.URN{failingResp.URN},
-		})
-		assert.NoError(t, err)
-
-		return nil
-	})
-	hostF := deploytest.NewPluginHostF(nil, nil, programF, loaders...)
-
-	p := &TestPlan{
-		Options: TestUpdateOptions{
-			UpdateOptions: UpdateOptions{
-				ContinueOnError: true,
-			},
-			HostF: hostF,
-		},
-	}
-
-	project := p.GetProject()
-
-	// Run an update to create the resource
-	snap, err := TestOp(Update).Run(project, p.GetTarget(t, nil), p.Options, false, p.BackendClient, nil)
-	require.ErrorContains(t, err, "intentionally failed create")
-	require.NotNil(t, snap)
-	require.Equal(t, 5, len(snap.Resources)) // 3 resources + 2 providers
-	assert.Equal(t, resource.URN("urn:pulumi:test::test::pulumi:providers:pkgB::default"), snap.Resources[0].URN)
-	assert.Equal(t, resource.URN("urn:pulumi:test::test::pulumi:providers:pkgA::default"), snap.Resources[1].URN)
-	assert.Equal(t, resource.URN("urn:pulumi:test::test::pkgA:m:typA::independent1"), snap.Resources[2].URN)
-	assert.Equal(t, resource.URN("urn:pulumi:test::test::pkgA:m:typA::independent2"), snap.Resources[3].URN)
-	assert.Equal(t, resource.URN("urn:pulumi:test::test::pkgA:m:typA::independent3"), snap.Resources[4].URN)
-}
-
-func TestUpContinueOnErrorUpdate(t *testing.T) {
-	t.Parallel()
-
-	loaders := []*deploytest.ProviderLoader{
-		deploytest.NewProviderLoader("pkgA", semver.MustParse("1.0.0"), func() (plugin.Provider, error) {
-			return &deploytest.Provider{}, nil
-		}, deploytest.WithoutGrpc),
-		deploytest.NewProviderLoader("pkgB", semver.MustParse("1.0.0"), func() (plugin.Provider, error) {
-			return &deploytest.Provider{
-				UpdateF: func(urn resource.URN, id resource.ID,
-					oldInputs, oldOutputs, newInputs resource.PropertyMap,
-					timeout float64, ignoreChanges []string, preview bool,
-				) (resource.PropertyMap, resource.Status, error) {
-					return nil, resource.StatusOK, errors.New("intentionally failed update")
-				},
-			}, nil
-		}, deploytest.WithoutGrpc),
-	}
-
-	ins := resource.NewPropertyMapFromMap(map[string]interface{}{
-		"foo": "bar",
-	})
-
-	// update as opposed to create for the failing resource
-	update := false
-
-	programF := deploytest.NewLanguageRuntimeF(func(_ plugin.RunInfo, monitor *deploytest.ResourceMonitor) error {
-		_, err := monitor.RegisterResource("pkgB:m:typB", "failing", true, deploytest.ResourceOptions{
-			SupportsResultReporting: true,
-			Inputs:                  ins,
-		})
-		assert.NoError(t, err)
-
-		if update {
-			respIndependent1, err := monitor.RegisterResource(
-				"pkgA:m:typA", "independent1", true, deploytest.ResourceOptions{
-					SupportsResultReporting: true,
-				})
-			assert.NoError(t, err)
-			respIndependent2, err := monitor.RegisterResource(
-				"pkgA:m:typA", "independent2", true, deploytest.ResourceOptions{
-					SupportsResultReporting: true,
-					Dependencies:            []resource.URN{respIndependent1.URN},
-				})
-			assert.NoError(t, err)
-			_, err = monitor.RegisterResource("pkgA:m:typA", "independent3", true, deploytest.ResourceOptions{
-				SupportsResultReporting: true,
-				Dependencies:            []resource.URN{respIndependent2.URN},
-			})
-			assert.NoError(t, err)
-		}
-
-		return nil
-	})
-
-	hostF := deploytest.NewPluginHostF(nil, nil, programF, loaders...)
-
-	p := &TestPlan{
-		Options: TestUpdateOptions{
-			UpdateOptions: UpdateOptions{
-				ContinueOnError: true,
-			},
-			HostF: hostF,
-		},
-	}
-
-	project := p.GetProject()
-
-	// Run an update to create the resource
-	snap, err := TestOp(Update).Run(project, p.GetTarget(t, nil), p.Options, false, p.BackendClient, nil)
-	require.NoError(t, err)
-	assert.NotNil(t, snap)
-	assert.Equal(t, 2, len(snap.Resources)) // 1 resource + 1 provider
-
-	update = true
-	ins = resource.NewPropertyMapFromMap(map[string]interface{}{
-		"foo": "baz",
-	})
-	// Run an update to create the resource
-	snap, err = TestOp(Update).Run(project, p.GetTarget(t, snap), p.Options, false, p.BackendClient, nil)
-	require.ErrorContains(t, err, "intentionally failed update")
-	assert.NotNil(t, snap)
-	assert.Equal(t, 6, len(snap.Resources)) // 4 resources + 2 providers
-	assert.Equal(t, resource.URN("urn:pulumi:test::test::pulumi:providers:pkgB::default"), snap.Resources[0].URN)
-	assert.Equal(t, resource.URN("urn:pulumi:test::test::pulumi:providers:pkgA::default"), snap.Resources[1].URN)
-	assert.Equal(t, resource.URN("urn:pulumi:test::test::pkgA:m:typA::independent1"), snap.Resources[2].URN)
-	assert.Equal(t, resource.URN("urn:pulumi:test::test::pkgA:m:typA::independent2"), snap.Resources[3].URN)
-	assert.Equal(t, resource.URN("urn:pulumi:test::test::pkgA:m:typA::independent3"), snap.Resources[4].URN)
-	assert.Equal(t, resource.URN("urn:pulumi:test::test::pkgB:m:typB::failing"), snap.Resources[5].URN)
-	assert.Equal(t, resource.NewStringProperty("bar"), snap.Resources[5].Inputs["foo"])
-}
-
 // TestStackOutputsResourceError tests that previous stack outputs aren't deleted when an update fails
 // due to a resource operation error.
 func TestStackOutputsResourceError(t *testing.T) {
